# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a HR management system API.",
        "title": "HR Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/corehr/assets": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of all assets with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get all assets",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Asset"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new asset with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Create a new asset",
                "parameters": [
                    {
                        "description": "Asset object",
                        "name": "asset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Asset"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/corehr/assets/serial/{serialNumber}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get detailed information about an asset by its serial number",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get asset by serial number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Serial Number",
                        "name": "serialNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Asset"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/corehr/assets/status/{status}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of assets with a specific status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get assets by status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status (available, assigned, maintenance, retired)",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Asset"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/corehr/assets/tag/{assetTag}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get detailed information about an asset by its asset tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get asset by asset tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Tag",
                        "name": "assetTag",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Asset"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/corehr/assets/type/{type}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of assets of a specific type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get assets by type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Type (laptop, phone, tablet, etc.)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Asset"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/corehr/assets/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get detailed information about an asset",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get asset by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Asset"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing asset's details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Update an asset",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Asset update object",
                        "name": "asset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Asset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete an asset by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Delete an asset",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_thecemakin_hr-project_internal_modules_corehr_model.Asset": {
            "type": "object",
            "properties": {
                "asset_tag": {
                    "description": "Internal tracking number",
                    "type": "string"
                },
                "assignments": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.AssetAssignment"
                    }
                },
                "brand": {
                    "type": "string"
                },
                "condition": {
                    "description": "new, good, fair, poor",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "description": "e.g., \"MacBook Pro 16-inch\"",
                    "type": "string"
                },
                "purchase_date": {
                    "type": "string"
                },
                "purchase_price": {
                    "description": "Purchase price in dollars",
                    "type": "number"
                },
                "serial_number": {
                    "type": "string"
                },
                "status": {
                    "description": "available, assigned, maintenance, retired",
                    "type": "string"
                },
                "type": {
                    "description": "laptop, phone, tablet, etc.",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "warranty_expires": {
                    "type": "string"
                }
            }
        },
        "github_com_thecemakin_hr-project_internal_modules_corehr_model.AssetAssignment": {
            "type": "object",
            "properties": {
                "asset": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Asset"
                        }
                    ]
                },
                "asset_id": {
                    "type": "integer"
                },
                "assigned_by": {
                    "description": "ID of the person who assigned the asset",
                    "type": "integer"
                },
                "assigned_date": {
                    "type": "string"
                },
                "assigner": {
                    "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "employee": {
                    "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee"
                },
                "employee_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "returned_by": {
                    "description": "ID of the person who returned the asset",
                    "type": "integer"
                },
                "returned_date": {
                    "type": "string"
                },
                "returner": {
                    "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee"
                },
                "status": {
                    "description": "Status indicates current state of the assignment",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee": {
            "type": "object",
            "properties": {
                "address_line1": {
                    "description": "Address Information",
                    "type": "string"
                },
                "address_line2": {
                    "type": "string"
                },
                "bank_account_number": {
                    "description": "Bank Account Details",
                    "type": "string"
                },
                "bank_name": {
                    "type": "string"
                },
                "bank_routing_number": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "department": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Department"
                        }
                    ]
                },
                "department_id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "emergency_contact_name": {
                    "description": "Emergency Contact",
                    "type": "string"
                },
                "emergency_contact_phone": {
                    "type": "string"
                },
                "emergency_contact_relation": {
                    "type": "string"
                },
                "employee_number": {
                    "description": "Employment Information",
                    "type": "string"
                },
                "first_name": {
                    "description": "Personal Information",
                    "type": "string"
                },
                "gender": {
                    "description": "male, female, other",
                    "type": "string"
                },
                "hire_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "manager": {
                    "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee"
                },
                "manager_id": {
                    "description": "Self-referencing foreign key",
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Position"
                },
                "position_id": {
                    "type": "integer"
                },
                "postal_code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "description": "active, inactive, terminated",
                    "type": "string"
                },
                "subordinates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_thecemakin_hr-project_internal_modules_corehr_model.Employee"
                    }
                },
                "termination_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer \u003cyour-jwt-token\u003e\" to authenticate.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}